Move bindings code to driver crate.
So that the webassembly bindings are now just a tiny wrapper for the
driver crate.
